Integer

+++ src/OpenTTY.java
@@ -216,7 +216,7 @@ public class OpenTTY extends MIDlet implements CommandListener {
         return signal;
     }
     // | (User Manager)
-    public int getUserID(String user) { return user.equals("root") ? 0 : user.equals(username) ? 1000 : userID.containsKey(user) ? (Integer) userID.get(user) : -1; }
+    public int getUserID(String user) { return user.equals("root") ? 0 : user.equals(username) ? 1000 : userID.containsKey(user) ? ((Integer) userID.get(user)).intValue() : -1; }
     // | (Trackers)
     public Hashtable getprocess(String pid) { return sys.containsKey(pid) ? (Hashtable) sys.get(pid) : null; }
     public Object getobject(String pid, String item) { return sys.containsKey(pid) ? ((Hashtable) sys.get(pid)).get(item) : null; }
